the qin2440 branch compiled ok

+++ arch/arm/mach-qin2440/map.c
@@ -1,150 +1,146 @@
-#include <asm-arm/page.h>
 #include <linux/kernel.h>
 #include <linux/init.h>
-#include <asm-arm/mach/map.h>
-#include <asm-arm/memory.h>
-#include <asm-arm/sizes.h>
-#include <asm-arm/io.h>
 #include <mach/hardware.h>
 #include <mach/virt_addr.h>
+#include <asm/mach/map.h>
 
 static struct map_desc qin2440_iodesc[] __initdata = {
 	[0] = {
 		/* the BWSCON register */
-		.virtual = __BWSCON,
+		.virtual = (unsigned int)__BWSCON,
 		.pfn     = __phys_to_pfn(BWSCON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[1] = {
 		/* USB Host Controller */
-		.virtual = __HcRevision,
+		.virtual = (unsigned int)__HcRevision,
 		.pfn     = __phys_to_pfn(HcRevision),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[2] = {
 		/* Interrupt Controller */
-		.virtual = __SRCPND,
+		.virtual = (unsigned int)__SRCPND,
 		.pfn     = __phys_to_pfn(SRCPND),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[3] = {
 		/* DMA */
-		.virtual = __DISRC0,
+		.virtual = (unsigned int)__DISRC0,
 		.pfn     = __phys_to_pfn(DISRC0),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[4] = {
 		/* Clock & Power Management */
-		.virtual = __LOCKTIME,
+		.virtual = (unsigned int)__LOCKTIME,
 		.pfn     = __phys_to_pfn(LOCKTIME),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[5] = {
 		/* LCD Controller */
-		.virtual = __LCDCON1,
+		.virtual = (unsigned int)__LCDCON1,
 		.pfn     = __phys_to_pfn(LCDCON1),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[6] = {
 		/* NAND Flash */
-		.virtual = __NFCONF,
+		.virtual = (unsigned int)__NFCONF,
 		.pfn     = __phys_to_pfn(NFCONF),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[7] = {
 		/* Camera Interface */
-		.virtual = __CISRCFMT,
+		.virtual = (unsigned int)__CISRCFMT,
 		.pfn     = __phys_to_pfn(CISRCFMT),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[8] = {
 		/* UART */
-		.virtual = __ULCON0,
+		.virtual = (unsigned int)__ULCON0,
 		.pfn     = __phys_to_pfn(ULCON0),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[9] = {
 		/* PWM Timer */
-		.virtual = __TCFG0,
+		.virtual = (unsigned int)__TCFG0,
 		.pfn     = __phys_to_pfn(TCFG0),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[10] = {
 		/* USB Device */
-		.virtual = __FUNC_ADDR_REG,
+		.virtual = (unsigned int)__FUNC_ADDR_REG,
 		.pfn     = __phys_to_pfn(FUNC_ADDR_REG),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[11] = {
 		/* Watchdog Timer */
-		.virtual = __WTCON,
+		.virtual = (unsigned int)__WTCON,
 		.pfn     = __phys_to_pfn(WTCON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[12] = {
 		/* IIC */
-		.virtual = __IICCON,
+		.virtual = (unsigned int)__IICCON,
 		.pfn     = __phys_to_pfn(IICCON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[13] = {
 		/* IIS */
-		.virtual = __IISCON,
+		.virtual = (unsigned int)__IISCON,
 		.pfn     = __phys_to_pfn(IISCON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[14] = {
 		/* I/O port */
-		.virtual = __GPACON,
+		.virtual = (unsigned int)__GPACON,
 		.pfn     = __phys_to_pfn(GPACON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[15] = {
 		/* RTC */
-		.virtual = __RTCCON,
+		.virtual = (unsigned int)__RTCCON,
 		.pfn     = __phys_to_pfn(RTCCON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[16] = {
 		/* A/D Converter */
-		.virtual = __ADCCON,
+		.virtual = (unsigned int)__ADCCON,
 		.pfn     = __phys_to_pfn(ADCCON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[17] = {
 		/* SPI */
-		.virtual = __SPCON0,
+		.virtual = (unsigned int)__SPCON0,
 		.pfn     = __phys_to_pfn(SPCON0),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[18] = {
 		/* SD Interface */
-		.virtual = __SDICON,
+		.virtual = (unsigned int)__SDICON,
 		.pfn     = __phys_to_pfn(SDICON),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
 	},
 	[19] = {
 		/* AC97 Audio-CODEC Interface */
-		.virtual = __AC_GLBCTRL,
+		.virtual = (unsigned int)__AC_GLBCTRL,
 		.pfn     = __phys_to_pfn(AC_GLBCTRL),
 		.length  = SZ_512K,
 		.type    = MT_DEVICE,
